requirements.txt
fastapi
uvicorn
sqlalchemy
databases
alembic
python-jose
passlib[bcrypt]
httpx
pytest
pydantic
python-dotenv
mysql-connector-python
